fix some type errors
